Refactor some (unrelated) code into misc.c, reducing code duplication.
