+++ contrib/mod_sftp/misc.h
@@ -1,6 +1,6 @@
 /*
  * ProFTPD - mod_sftp miscellaneous
- * Copyright (c) 2010-2015 TJ Saunders
+ * Copyright (c) 2010-2016 TJ Saunders
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -29,5 +29,6 @@
 
 int sftp_misc_chown_file(pool *, pr_fh_t *);
 int sftp_misc_chown_path(pool *, const char *);
+const char *sftp_misc_namelist_shared(pool *, const char *, const char *);
 
 #endif /* MOD_SFTP_MISC_H */
